+++ src/lib/utils/domUtils.ts
@@ -66,7 +66,10 @@ export function jsonEscapeUnicode(value: string): string {
 export function jsonUnescapeUnicode(value: string): string {
   return value.replace(/\\u[a-fA-F0-9]{4}/g, (x) => {
     try {
-      return JSON.parse('"' + x + '"')
+      const unescaped = JSON.parse('"' + x + '"')
+      // the resolved character can be a control character like " or \n,
+      // that would result in invalid JSON, so we need to keep that escaped
+      return controlCharacters[unescaped] || unescaped
     } catch (err) {
       return x
     }
@@ -74,6 +77,7 @@ export function jsonUnescapeUnicode(value: string): string {
 }
 
 const controlCharacters = {
+  '"': '\\"',
   '\\': '\\\\',
   // escaped forward slash '\/' is the same as '/', we can't escape/unescape it
   '\b': '\\b',
@@ -85,6 +89,7 @@ const controlCharacters = {
 }
 
 const escapedControlCharacters = {
+  '\\"': '"',
   '\\\\': '\\',
   // escaped forward slash '\/' is the same as '/', we can't escape/unescape it
   '\\/': '/',
@@ -97,13 +102,13 @@ const escapedControlCharacters = {
 }
 
 export function jsonEscapeControl(value: string): string {
-  return value.replace(/[\b\f\n\r\t\\]/g, (x) => {
+  return value.replace(/["\b\f\n\r\t\\]/g, (x) => {
     return controlCharacters[x] || x
   })
 }
 
 export function jsonUnescapeControl(value: string): string {
-  return value.replace(/\\[bfnrt\\]/g, (x) => {
+  return value.replace(/\\["bfnrt\\]/g, (x) => {
     return escapedControlCharacters[x] || x
   })
 }